Big Upgrade
Fixed Racer Tags in freeroam - they will not show up even if you have them checked now.
Added the scroll to the lava in the dance scene.
Added some preventative measure if you massively widen or shorten the window for the dance scene - now it just cuts it off.
Fixed the credits menu with that bizarre window issue
Fixed the audio slider in the options menu
Replaced the old event system on some of the menus
Updated some of the menu canvas to 1920 by 1080, still have more work for that
Added another skin for the dreadstrider and another for speedstinger
fixed a legacy input field on the store